Added checking for username availability

+++ server/router.ts
@@ -1,13 +1,18 @@
 import express, { Request, Response } from 'express'
 import fs from 'fs'
 import { v4 as uuidv4 } from 'uuid'
+import { PrismaClient } from '@prisma/client'
+
+const prisma = new PrismaClient()
 
 const router = express.Router()
 
-router.post('/auth', (req: Request, res: Response) => {
+router.post('/auth', async (req: Request, res: Response) => {
   const { actionType, email, password, username, image } = req.body
   if (actionType === 'register') {
     if (!email || !password || !username || !image) return res.status(400).json({ error: 'Bad Request' })
+    const existingUser = await prisma.user.findFirst({ where: { username } })
+    if (existingUser) return res.status(409).json({ error: 'Username in use' })
     const base64Image: string = image.split(';base64,').pop()
     const imageUrl = `assets/profilePictures/${uuidv4()}.png`
     fs.writeFileSync(imageUrl, base64Image, { encoding: 'base64' })
